backend/requirement.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
httpx==0.28.1

# ── LangChain ecosystem ────────────────────────────────────────────────────────
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.18
langchain-text-splitters==1.0.0
langchain-google-genai==3.0.2
langchain-ollama==0.3.10
langchain-qdrant==1.1.0

# ── Vector DB ─────────────────────────────────────────────────────────────────
fastembed==0.7.3
qdrant-client==1.15.1

# ── Reranking ─────────────────────────────────────────────────────────────────
FlashRank

# ── LLM providers ─────────────────────────────────────────────────────────────
google-genai==1.38.0

# ── Document parsing ──────────────────────────────────────────────────────────
python-docx==1.2.0
PyPDF2==3.0.1

# ── Database ──────────────────────────────────────────────────────────────────
beanie==1.26.0
motor==3.7.1

# ── Auth / Pydantic ───────────────────────────────────────────────────────────
PyJWT==2.10.1
pydantic==2.12.0
colorama==0.4.6
